Aborted making BV-EOF vectors being saved with underlying velocity flow added since it didnt work and I wouldnt use more time on it. Cleaned a bit up in the TL and ATL setup in the perturbation_runner.py. Implemented Numba for the ATL model. Removed du_array from all L63 models since it wasnt needed. Added ref_data to a local argument since it wasnt needed outside the atlm_model.

+++ general/utils/importing/import_perturbation_data.py
@@ -292,7 +292,10 @@ def import_perturb_vectors(
     vector_units = np.empty(
         (args["n_files"], args["n_runs_per_profile"], params.sdim), dtype=np.float64
     )
-    characteristic_values = np.empty((args["n_files"], args["n_runs_per_profile"]))
+
+    characteristic_values = np.empty(
+        (args["n_files"], args["n_runs_per_profile"]), dtype=np.complex128
+    )
 
     for i, file_name in enumerate(perturb_file_names):
         vector_unit, _ = g_import.import_data(
